vault-desktop/requirements.txt
fastmcp>=0.9.0
httpx>=0.25.0
pystray>=0.19.4
requests-oauthlib>=1.3.1
keyring>=24.0.0
pydantic>=2.0.0
pillow>=10.0.0
asyncio>=3.4.3
ollama>=0.3.0